\name{diam}
\alias{diam}
\docType{data}
\title{
diameter data from an AgCl colloid with gnerally cubic particles.
}
\description{
A dataframe containing the equivalent circular diameter (in nm) measured by image analysis in the transmission electron microscope with the sample cooled to -170C to prevent beam-induced reduction. 
}
\usage{data(diam)}
\format{
  A data frame with 2123 observations on the following variable.
  \describe{
    \item{\code{ecd.nm}}{a numeric vector}
  }
}
\details{
An AgCl colloid with gnerally cubic particles.
}
\source{
Unpublished data
}
\references{
Unpublished
}
\examples{
data(diam)
## maybe str(diam) ; plot(diam) ...
}
\keyword{datasets}
\keyword{diameter}
